Bug 1596513: Part 2: Allow for floating point dx/dy on feDropShadow attributes r=nical
WebRender expects drop shadow dx/dy to be in user space, which is a floating point value.
Before this, a dx/dy such as (0.2, 0.2) was truncated to (0, 0) resulting in no offset of the shadow.

Differential Revision: https://phabricator.services.mozilla.com/D63442

+++ layout/svg/nsFilterInstance.cpp
@@ -257,7 +257,7 @@ bool nsFilterInstance::BuildWebRenderFilters(nsIFrame* aFilteredFrame,
                       gsRGBToLinearRGBMap[uint8_t(color.b * 255)], color.a);
       }
       wr::Shadow wrShadow;
-      wrShadow.offset = {(float)shadow.mOffset.x, (float)shadow.mOffset.y};
+      wrShadow.offset = {shadow.mOffset.x, shadow.mOffset.y};
       wrShadow.color = wr::ToColorF(ToDeviceColor(color));
       wrShadow.blur_radius = stdDev.width;
       wr::FilterOp filterOp = wr::FilterOp::DropShadow(wrShadow);
